chore: update `time` provider to `~> 0.7` (#80)
* fix: update the time provider version allowed
Hashicorp's `time` provider 0.6.x does not have a binary for the
darwin_arm64 platform, so developers on the new M1/M2 Macs
will not be able to apply this module. The 0.7.x release includes
a darwin_arm64 binary and makes no other changes.
I would also note that `time` is a very stable provider overall
(current version 0.9.1, released Nov 2022). I recommend switching
any `~>` constraints to `>=` unless there is a documented reason
not to.

* chore: fmt
* docs(readme): keep readme up to date with code